More merge changes.
